+++ lambdas/data-migration1/src/index.ts
@@ -2,7 +2,6 @@ import { getKnexClient } from '@cumulus/db';
 import { MigrationSummary } from '@cumulus/types/migration';
 import Logger from '@cumulus/logger';
 
-import { migrateCollections } from './collections';
 import { migrateProviders } from './providers';
 import { migrateAsyncOperations } from './async-operations';
 import { migrateRules } from './rules';
@@ -17,19 +16,12 @@ export const handler = async (event: HandlerEvent): Promise<MigrationSummary> =>
   const knex = await getKnexClient({ env });
 
   try {
-    const collectionsMigrationSummary = await migrateCollections(env, knex);
     const providersMigrationSummary = await migrateProviders(env, knex);
     const asyncOpsMigrationSummary = await migrateAsyncOperations(env, knex);
     const rulesMigrationSummary = await migrateRules(env, knex);
 
     const result: MigrationSummary = {
       MigrationSummary: {
-        collections: {
-          total_dynamo_db_records: collectionsMigrationSummary.dynamoRecords,
-          migrated: collectionsMigrationSummary.success,
-          skipped: collectionsMigrationSummary.skipped,
-          failed: collectionsMigrationSummary.failed,
-        },
         providers: {
           total_dynamo_db_records: providersMigrationSummary.dynamoRecords,
           migrated: providersMigrationSummary.success,